AB#8 First refactor of skeletal idea
